backend/requirements.txt
Django==5.2.6
django-ninja==1.4.3
djangorestframework==3.16.1
requests==2.32.5
beautifulsoup4==4.13.5
gunicorn==23.0.0
pillow==11.3.0
ruff==0.14.0
coverage==7.10.7
pydantic==2.11.10
